rename \Gems\Events to \Gems\Tracker\TrackEvents and add to legacy config provider
